refactor: correcting faults in case of use

+++ src/trajectory/src/domain/src/main/java/com/buildingBlocks/trajectory/domain/player/values/Diner.java
@@ -4,14 +4,14 @@ import com.buildingblocks.shared.domain.generic.IValueObject;
 import com.buildingblocks.shared.domain.utils.ValidateUtils;
 
 public class Diner implements IValueObject {
-    private final long amount;
+    private final Integer amount;
 
-    public Diner(long amount){
+    public Diner(Integer amount){
         this.amount = amount;
         validate();
     }
 
-    public static Diner of(long amount){
+    public static Diner of(Integer amount){
         return new Diner(amount);
     }
 
@@ -21,7 +21,7 @@ public class Diner implements IValueObject {
         ValidateUtils.validateMaxLength(String.valueOf(amount), 1000, "el dinero no puede ser mayor a 1000");
     }
 
-    public long getAmount() {
+    public Integer getAmount() {
         return amount;
     }
 }
